Make note metadata contain itself
This allows the body part of the "note" metadata to generate itself, or
modified versions of itself.
This fixes the hacky solution from before where the body contained a
non-root copy of itself, which then contained a metadata with nothing in
the body. There is no longer a hacky depth limit.