Change license to GPLv3
In a perfect world everyone would use AGPLv3, but this is not where we
are. In order for GPLv3 projects to make use of this library, it cannot
be AGPL. And in reality the extra provisions of the AGPL don't actually
do anything for this library. So, GPLv3 it is.